+++ build_signaler/build_signaler.go
@@ -1,10 +1,11 @@
 package build_signaler
 
 import (
-	"errors"
 	"fmt"
+	"strings"
 	"time"
 
+	"github.com/pkg/errors"
 	"github.com/shankj3/go-til/deserialize"
 	"github.com/shankj3/go-til/log"
 	ocenet "github.com/shankj3/go-til/net"
@@ -12,6 +13,8 @@ import (
 	"github.com/shankj3/ocelot/models/pb"
 	"github.com/shankj3/ocelot/storage"
 )
+
+const SubscriptionUpdateBranch = "master"
 
 func storeStageToDb(store storage.BuildStage, stageResult *models.StageResult) error {
 	if err := store.AddStageDetail(stageResult); err != nil {
@@ -38,8 +41,6 @@ func storeSummaryToDb(store storage.BuildSum, hash, repo, branch, account string
 	return id, nil
 }
 
-//todo: pull out check for vcsHandler == nil logic, then this can be just GetConfig()
-// todo (cont): write something in remote to switch between subtypes to instantiate the correct VCSHandler implementation
 //GetConfig returns the protobuf ocelot.yaml, a valid bitbucket token belonging to that repo, and possible err.
 //If a VcsHandler is passed, this method will use the existing handler to retrieve the bb config. In that case,
 //***IT WILL NOT RETURN A VALID TOKEN FOR YOU - ONLY BUILD CONFIG***
@@ -60,6 +61,36 @@ func GetConfig(repoFullName string, checkoutCommit string, deserializer *deseria
 	conf, err := CheckForBuildFile(fileBytz, deserializer)
 	return conf, err
 }
+
+// CreateOrUpdateSubscription will build out an ActiveSubscription model and then create or update the the subscription in storage.
+func CreateOrUpdateSubscription(conf *pb.BuildConfig, signal *Signaler, subscribingAcctRepo string, subscribingVcsType pb.SubCredType) error {
+	for _, subscription := range conf.GetSubscriptions() {
+		sub := &pb.ActiveSubscription{
+			BranchQueueMap: make(map[string]string),
+			SubscribedToAcctRepo: subscription.AcctRepo,
+			SubscribedToVcsType: subscription.AcctVcsType,
+			SubscribingVcsType: subscribingVcsType,
+			SubscribingAcctRepo: subscribingAcctRepo,
+			Alias: subscription.Alias,
+		}
+		for _, branchmap := range subscription.GetBranches() {
+			split := strings.Split(branchmap, ":")
+			if len(split) != 2 {
+				return errors.New("unparseable branch mapping")
+			}
+			upstreamBranch, subscribingBranch := split[0], split[1]
+			sub.BranchQueueMap[upstreamBranch] = subscribingBranch
+		}
+		id, err := signal.Store.InsertOrUpdateActiveSubscription(sub)
+		if err != nil {
+			log.IncludeErrField(err).WithField("ActiveSubscription", sub).Error("unable to insert active subscription!")
+			return errors.Wrap(err, "unable to insert active subscription")
+		}
+		log.Log().WithField("ActiveSubscription", sub).Infof("successful insert %d", id)
+	}
+	return nil
+}
+
 
 //CheckForBuildFile will try to retrieve an ocelot.yaml file for a repository and return the protobuf message
 func CheckForBuildFile(buildFile []byte, deserializer *deserialize.Deserializer) (*pb.BuildConfig, error) {
