+++ build_signaler/push_werkerteller.go
@@ -32,6 +32,16 @@ func (pwt *PushWerkerTeller) TellWerker(push *pb.Push, conf *Signaler, handler m
 		ocelog.IncludeErrField(err).Error("couldn't get ocelot.yml")
 		return errors.Wrap(err, "unable to get build configuration")
 	}
+
+	if push.Branch == SubscriptionUpdateBranch && buildConf.GetSubscriptions() != nil {
+		ocelog.Log().WithField("hash", push.HeadCommit.Hash).WithField("acctRepo", push.Repo.AcctRepo).Info("inserting or updating subscription")
+		if err = CreateOrUpdateSubscription(buildConf, conf, push.Repo.AcctRepo, handler.GetVcsType()); err != nil {
+			ocelog.IncludeErrField(err).WithField("hash", push.HeadCommit.Hash).WithField("acctRepo", push.Repo.AcctRepo).Error("unable to update subscriptions table")
+			// todo: not sure if we should return if this fails... seems like we should still want to build.
+		} else {
+			ocelog.Log().WithField("hash", push.HeadCommit.Hash).WithField("acctRepo", push.Repo.AcctRepo).Info("successfully inserted/updated subscription")
+		}
+	}
 	task := BuildInitialWerkerTask(buildConf, push.HeadCommit.Hash, token, push.Branch, push.Repo.AcctRepo, sigBy, nil, handler.GetVcsType())
 	if push.PreviousHeadCommit == nil {
