+++ common/trigger/conditiontypes.go
@@ -157,3 +157,40 @@ func (b *FilepathCondition) AddConditionValue(str string) {
 func (b *FilepathCondition) GetConditionValues() []string {
 	return b.acceptedFilepaths
 }
+
+
+type UpstreamCondition struct {
+	goodUpstreamAliases []string
+	logical 			Conditional
+}
+
+func (b *UpstreamCondition) GetTriggerType() TriggerType {
+	return Upstream
+}
+
+// PassesMuster makes sure that the alias of the subscription that triggered the active build is in one of the
+//  accepted aliasses denoted by \upstream alias1||alias2||alias3 (etc)
+func (b *UpstreamCondition) PassesMuster(td *pb.ChangesetData) bool {
+	for _, acceptedAlias := range b.goodUpstreamAliases {
+		if acceptedAlias == td.SubscriptionAlias {
+			return true
+		}
+	}
+	return false
+}
+
+func (b *UpstreamCondition) GetLogical() Conditional {
+	return b.logical
+}
+
+func (b *UpstreamCondition) SetLogical(conditional Conditional) {
+	b.logical = conditional
+}
+
+func (b *UpstreamCondition) AddConditionValue(str string) {
+	b.goodUpstreamAliases = append(b.goodUpstreamAliases, str)
+}
+
+func (b *UpstreamCondition) GetConditionValues() []string {
+	return b.goodUpstreamAliases
+}
