+++ common/trigger/types.go
@@ -11,6 +11,7 @@ const (
 	Branch
 	Filepath
 	Text
+	Upstream
 )
 
 func (t TriggerType) Spawn() Section {
@@ -21,6 +22,8 @@ func (t TriggerType) Spawn() Section {
 		return &FilepathCondition{}
 	case Text:
 		return &TextCondition{}
+	case Upstream:
+		return &UpstreamCondition{}
 	default:
 		panic("shouldn't happen")
 	}
@@ -34,6 +37,8 @@ func convertTriggerType(str string) TriggerType {
 		return Filepath
 	case `\text`:
 		return Text
+	case `\upstream`:
+		return Upstream
 	default:
 		return TNone
 	}
