+++ go.mod
@@ -2,15 +2,18 @@ module github.com/shankj3/ocelot
 
 require (
 	code.cloudfoundry.org/bytefmt v0.0.0-20180906201452-2aa6f33b730c
+	github.com/NYTimes/gziphandler v1.0.1 // indirect
 	github.com/docker/distribution v2.6.2+incompatible // indirect
 	github.com/docker/docker v1.13.1
 	github.com/docker/go-connections v0.4.0
+	github.com/elazarl/go-bindata-assetfs v1.0.0 // indirect
 	github.com/fatih/color v1.7.0
+	github.com/go-sql-driver/mysql v1.4.0 // indirect
 	github.com/go-test/deep v1.0.1
 	github.com/golang/mock v1.1.1
 	github.com/golang/protobuf v1.2.0
 	github.com/google/go-github/v19 v19.1.0
-	github.com/google/uuid v1.1.0
+	github.com/google/uuid v1.0.0
 	github.com/gorhill/cronexpr v0.0.0-20180427100037-88b0669f7d75
 	github.com/gorilla/context v1.1.1 // indirect
 	github.com/gorilla/mux v1.6.2
@@ -18,31 +21,32 @@ require (
 	github.com/grpc-ecosystem/go-grpc-prometheus v1.2.0
 	github.com/grpc-ecosystem/grpc-gateway v1.5.1
 	github.com/hashicorp/consul v1.4.0
+	github.com/hashicorp/hcl v1.0.0 // indirect
+	github.com/hashicorp/serf v0.8.1 // indirect
 	github.com/hashicorp/vault v1.0.0-rc1
-	github.com/lib/pq v1.0.0
+	github.com/lib/pq v0.0.0-20181016162627-9eb73efc1fcc
+	github.com/mattn/go-isatty v0.0.4 // indirect
 	github.com/mattn/go-runewidth v0.0.3 // indirect
 	github.com/mitchellh/cli v1.0.0
 	github.com/namsral/flag v1.7.4-pre
 	github.com/olekukonko/tablewriter v0.0.1
+	github.com/patrickmn/go-cache v2.1.0+incompatible // indirect
 	github.com/philips/grpc-gateway-example v0.0.0-20170619012617-a269bcb5931c
 	github.com/pkg/errors v0.8.0
+	github.com/posener/complete v1.2.1 // indirect
 	github.com/prometheus/client_golang v0.9.1
 	github.com/prometheus/client_model v0.0.0-20180712105110-5c3871d89910
-	github.com/shankj3/go-til v0.4.3
+	github.com/prometheus/common v0.0.0-20181126121408-4724e9255275 // indirect
+	github.com/prometheus/procfs v0.0.0-20181005140218-185b4288413d // indirect
+	github.com/shankj3/go-til v0.4.3-1
+	github.com/sirupsen/logrus v1.2.0 // indirect
 	github.com/stevvooe/resumable v0.0.0-20180830230917-22b14a53ba50 // indirect
-<<<<<<< HEAD
-	golang.org/x/crypto v0.0.0-20181106171534-e4dc69e5b2fd
-	golang.org/x/net v0.0.0-20181106065722-10aee1819953
-	golang.org/x/oauth2 v0.0.0-20181106182150-f42d05182288
-	golang.org/x/sync v0.0.0-20181108010431-42b317875d0f // indirect
-	golang.org/x/sys v0.0.0-20181106135930-3a76605856fd // indirect
-	golang.org/x/tools v0.0.0-20181130052023-1c3d964395ce // indirect
-	google.golang.org/appengine v1.3.0 // indirect
-=======
 	golang.org/x/crypto v0.0.0-20181127143415-eb0de9b17e85
 	golang.org/x/net v0.0.0-20181129055619-fae4c4e3ad76
 	golang.org/x/oauth2 v0.0.0-20181128211412-28207608b838
->>>>>>> de017d2f272153639265c30eff7a363387f55173
+	golang.org/x/sync v0.0.0-20181108010431-42b317875d0f // indirect
+	golang.org/x/sys v0.0.0-20181106135930-3a76605856fd // indirect
+	google.golang.org/appengine v1.3.0 // indirect
 	google.golang.org/genproto v0.0.0-20181127195345-31ac5d88444a
 	google.golang.org/grpc v1.16.0
 	gopkg.in/yaml.v2 v2.2.2
